requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
sentence-transformers
faiss-cpu
groq
fastembed
python-multipart